Small refactoring/Preparation for summaries

+++ OctoAwesome/OctoAwesome.Network/Settings.cs
@@ -1,19 +1,15 @@
 ﻿//using OpenTK;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
-using System.Configuration;
 using System.IO;
 using System.Linq;
-using System.Reflection;
-using System.Runtime.Serialization;
 
 namespace OctoAwesome.Network
 {
     public class Settings : ISettings
     {
-        public FileInfo FileInfo { get; set; }
+        public FileInfo FileInfo { get; init; }
         private readonly Dictionary<string, string> dictionary;
 
         public Settings(FileInfo fileInfo)
@@ -21,6 +17,7 @@ namespace OctoAwesome.Network
             FileInfo = fileInfo;
             dictionary = InternalLoad(fileInfo);
         }
+
         public Settings()
         {
             dictionary = new Dictionary<string, string>()
@@ -30,28 +27,23 @@ namespace OctoAwesome.Network
                 ["DisablePersistence"] = "false",
                 ["LastUniverse"] = ""
             };
+            FileInfo = null!;
         }
-
         public void Delete(string key)
             => dictionary.Remove(key);
-
         public T Get<T>(string key)
             => (T)Convert.ChangeType(dictionary[key], typeof(T));
-
         public T Get<T>(string key, T defaultValue)
         {
-            if (dictionary.TryGetValue(key, out string value))
+            if (dictionary.TryGetValue(key, out var value))
                 return (T)Convert.ChangeType(value, typeof(T));
 
             return defaultValue;
         }
-
         public T[] GetArray<T>(string key)
             => DeserializeArray<T>(dictionary[key]);
-
         public bool KeyExists(string key)
             => dictionary.ContainsKey(key);
-
         public void Set(string key, string value)
         {
             if (dictionary.ContainsKey(key))
@@ -71,18 +63,21 @@ namespace OctoAwesome.Network
             => Set(key, values.Select(i => i.ToString()).ToArray());
         public void Set(string key, bool[] values)
             => Set(key, values.Select(b => b.ToString()).ToArray());
-
         public void Load()
         {
+            if (FileInfo == null)
+                throw new ArgumentException("No file info was specified");
+
             dictionary.Clear();
 
             foreach (var entry in InternalLoad(FileInfo))
                 dictionary.Add(entry.Key, entry.Value);
         }
-
-
         public void Save()
         {
+            if (FileInfo == null)
+                throw new ArgumentException("No file info was specified");
+
             FileInfo.Delete();
             using (var writer = new StreamWriter(FileInfo.OpenWrite()))
             {
@@ -100,10 +95,7 @@ namespace OctoAwesome.Network
 
         private T[] DeserializeArray<T>(string arrayString)
         {
-            // Wir müssten, um beide Klammern zu entfernen, - 3 rechnen. Ich lasse die letzte Klammer stellvertretend für das Komma, was folgen würde, stehen.
-            // Das wird in der for-Schleife auseinander gepflückt.
-
-            arrayString = arrayString.Substring(1, arrayString.Length - 2 /*- 1*/);
+            arrayString = arrayString.Substring(1, arrayString.Length - 2);
 
             string[] partsString = arrayString.Split(',');
             T[] tArray = new T[partsString.Length];
